updated loader and cinfirm and view dialog

+++ src/components/student-details-table/student-details-table.component.ts
@@ -1,6 +1,6 @@
 import { Component, inject, OnInit, ViewChild } from '@angular/core';
 import { StudentService } from '../../services/student.service';
-import {MatButtonModule} from '@angular/material/button';
+import { MatButtonModule } from '@angular/material/button';
 import { CommonModule } from '@angular/common';
 import { MatTableDataSource, MatTableModule } from '@angular/material/table';
 import { MatIconModule } from '@angular/material/icon';
@@ -13,6 +13,9 @@ import { MatPaginator, MatPaginatorIntl } from '@angular/material/paginator';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
 import { MatSort, Sort } from '@angular/material/sort';
 import { SnackbarService } from '../../services/snackbar.service';
+import { ConfirmDialogService } from '../../services/confirm-dialog.service';
+import { IConfirmDialogConfig } from '../../models/confirm-dialog';
+import { StudDetailViewDialogComponent } from '../stud-detail-view-dialog/stud-detail-view-dialog.component';
 
 @Component({
   selector: 'app-student-details-table',
@@ -27,28 +30,29 @@ export class StudentDetailsTableComponent implements OnInit {
 
 
 
-  displayedColumns: string[] = ['Sr No', 'firstName', 'lastName','action'];
+  displayedColumns: string[] = ['Sr No', 'firstName', 'lastName', 'action'];
   // dataSource: any[] = []; 
 
 
-  dataSource = new MatTableDataSource<IStudent>([]); 
+  dataSource = new MatTableDataSource<IStudent>([]);
   @ViewChild(MatPaginator) paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
 
-  itemsPerPageOptions: number[] = [5,10,15,20];
+  itemsPerPageOptions: number[] = [5, 10, 15, 20];
 
   constructor(
     private _studentService: StudentService,
     private _matDialog: MatDialog,
     private paginatorIntl: MatPaginatorIntl,
     private _snackBar: SnackbarService,
-  ) { 
+    private _confirmDialogService: ConfirmDialogService
+  ) {
     this.paginatorIntl.itemsPerPageLabel = 'Students per page:';
   }
 
   ngOnInit(): void {
     console.log('StudentDetailsTableComponent dataSource', this.dataSource);
-    
+
     this.getAllStudents();
     this._studentService.sendNewStudentAsObs$.subscribe((newStudent: any) => {
       this.dataSource.data = [...this.dataSource.data, newStudent];
@@ -57,14 +61,14 @@ export class StudentDetailsTableComponent implements OnInit {
       const index = this.dataSource.data.findIndex(student => student._id === updatedStudent._id);
       if (index !== -1) {
         this.dataSource.data[index] = updatedStudent;
-        this.dataSource.data = [...this.dataSource.data]; 
+        this.dataSource.data = [...this.dataSource.data];
 
       }
     });
   }
 
-   ngAfterViewInit() {
-        
+  ngAfterViewInit() {
+
     this.dataSource.paginator = this.paginator;
     this.dataSource.sort = this.sort;
   }
@@ -72,50 +76,94 @@ export class StudentDetailsTableComponent implements OnInit {
   getAllStudents() {
     this._studentService.getAllStudents().subscribe((students: any) => {
       console.log('Fetched students:', students.data);
-      this.dataSource.data = students.data; 
+      this.dataSource.data = students.data;
     }, error => {
-        console.error('Error fetching students:', error);
-    }); 
-  }
-
-  deleteStudent(studentId: any) {    
-    this._studentService.deleteStudent(studentId).subscribe({
-      next: (res:any) => {
-        if (res && res['success']) {
-         this._snackBar.show(res['message'], 'success');
-         this.getAllStudents(); 
-        }else {
-          this._snackBar.show(res['message'], 'warn');
-        }
-      },
-      error: (error) => {
-        console.error('Error deleting student:', error);
-      }
+      console.error('Error fetching students:', error);
     });
   }
 
-  updateStudent(studentId: any, studentData: any) {
-    this._studentService.isStudentFormEditMode$.next(true); 
+  deleteStudent(studentId: any) {
 
-   const dialogRef = this._matDialog.open(StudentRegFormComponent, {
+    const dialogConfig: IConfirmDialogConfig = {
+      title: 'Delete Student',
+      message: 'Are you sure you want to delete this student? ',
+      confirmText: 'Delete',
+      cancelText: 'Cancel',
+    }
+
+    this._confirmDialogService.openConfirmDialog(dialogConfig).subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        this._studentService.deleteStudent(studentId).subscribe({
+          next: (res: any) => {
+            if (res && res['success']) {
+              this._snackBar.show(res['message'], 'success');
+              this.getAllStudents();
+            } else {
+              this._snackBar.show(res['message'], 'warn');
+            }
+          },
+        });
+      } else {
+        this._snackBar.show('Student deletion cancelled', 'info');
+      }
+    });
+
+
+
+  }
+
+  updateStudent(studentId: any, studentData: any) {
+
+    const dialogConfig: IConfirmDialogConfig = {
+      title: 'Update Student',
+      message: `Are you sure you want to Update ${studentData.firstName} `,
+      confirmText: 'Yes',
+      cancelText: 'No',
+    }
+
+    this._confirmDialogService.openConfirmDialog(dialogConfig).subscribe((confirmed: boolean) => {
+      if (confirmed) {
+
+        this._studentService.isStudentFormEditMode$.next(true);
+
+        const dialogRef = this._matDialog.open(StudentRegFormComponent, {
+          width: '500px',
+          data: {
+            student: studentData
+          }
+        });
+        dialogRef.afterClosed().subscribe(() => {
+          this._studentService.isStudentFormEditMode$.next(false);
+        });
+      } else {
+        this._snackBar.show('Student update cancelled', 'info');
+        return;
+      }
+    });
+
+  }
+
+  viewStudent(student: any) {
+    
+    const dialogRef = this._matDialog.open(StudDetailViewDialogComponent, {
       width: '500px',
       data: {
-        student: studentData        
+        student: student,
+        isViewMode: true
       }
     });
     dialogRef.afterClosed().subscribe(() => {
       this._studentService.isStudentFormEditMode$.next(false);
     });
-
   }
 
   onSearchStudent(e: Event) {
-      const filterValue = (e.target as HTMLInputElement).value;
+    const filterValue = (e.target as HTMLInputElement).value;
     this.dataSource.filter = filterValue.trim().toLowerCase();
   }
 
   onSortChange(sortState: Sort) {
-    
+
     if (sortState.direction) {
       this._liveAnnouncer.announce(`Sorted ${sortState.direction}ending`);
     } else {
